source-backend: simplify and increase concurrency
Instead of trying to come up with some magic formula for how many
Goroutines make sense, we now exploit the fact that Goroutines are cheap
and rely on the Go runtime scheduler to make the best of the situation.

Additionally, this change should hide storage latency as good as
possible since all IO is being started at the same time.