add dcs-rs-deploy (main rackspace deployment automation)
As with the previous commit:
Note that there are still hard-coded paths to SSH keys in the code. In
case you are actually interested in using this for setting up your own
DCS instance on rackspace, let me know and we can make it work. Until
then, this is mostly for transparency and in case the code is useful for
anyone.