generate result pages as necessary
Previously, mostly due to the proof-of-concept design (collect all
results in memory, sort them, then write them out to disk and let nginx
serve them), we were writing the entire query to disk twice — once for
the regular results, once for per-package results.

For small queries, this is not a problem, but for big queries, that
consumes a significant amount of disk space and spindle time.

Therefore, we now generate (and cache!) the pages for which the
client(s) ask. This is a multi-minute speed-up for big queries.