delete dcs-batch-helper
The idea of it was to start a new batch instance for long/expensive
queries. At the moment, we can just handle these queries with our normal
stack, since we don’t get a lot of queries and the performance
degradation is not a problem anymore with the new architecture/hardware.