Strip leading/trailing filters, but preserve entire search term
Currently, extra whitespace around filters can cause a search to
incorrectly fail.  Whitespace within a search string is relevant, and
should be preserved, but whitespace between filters and the search
string should not be relevant.

That is " package:foo   bar    baz" is a search for "bar    baz" in
package foo, not a search for "  bar    baz".

This commit changes how filters are found by only looking for them at
the start/end of the search string.  Any found filters are stripped and
the remainder of the search string is used for the final query.

Signed-off-by: James McCoy <jamessan@debian.org>