results/placeholder: don’t include highlightjs
highlightjs is not actually used on the results pages, only on the show
pages. And even if we wanted to use it, the results template would be
the wrong place, as it’s only used when people don’t have javascript
enabled (with javascript enabled, you’re redirected).

saves 10KB, does not result in a wall-clock time page load improvement
(because the loading is parallelized).