Make headroom percentage configurable, disable in dcs-localdcs
This safe-guard is intended for the production installation. During
development, it’s fine to have less disk space. In case this leads to
errors, it would be good to fix them properly on a case-by-case basis
instead of relying on the headroom mechanism.

Notably, when the file system is full enough that there isn’t enough
headroom (without any queries being stored yet), the newly created
directory for the query which should be started will immediately be
deleted by the headroom mechanism.