TurboPFor: ensure input (!) is padded, too
We already ensured the output buffers we pass to TurboPFor are padded,
but it turns out that TurboPFor also reads past its input buffer!
(At least in the version we’re currently using.)

This manifested in crashes when merging multiple index files together,
and only first occurred starting a few weeks ago, presumably because a new
package entered the Debian archive that happens to trigger the bug.

Reading past the number of input bytes in the input buffer is not just an
invalid memory access, but a Segmentation Violation because we mmap
the input files.

fixes https://github.com/Debian/dcs/issues/123